no longer use vectors for html templating
The use of vector like `[a b c]` was problematic in one way
because it doesn't evaluate its arguments. `a`, `b` and `c`
are not evaluated in the form `[a b c]`.

The consequence is that we couldn't pass a variable in the template
system that make it almost useless.

Precisely:
  (let ((x "foo")) [x])
returns
  [x]
but I would have prefered ["foo"].

This led to:
  (let ((x "foo")) (osta-html [:p x]))
returning:
  "<p>x</p>"
when I would have prefered "<p>foo</p>".

Anyway, go back to list.

What we used to write
  [:p@id.class (:another-attr "foo") bar]
has been replaced by:
  '(:p:id.class (@ :another-attr "foo") bar)

And this allow us to write something like:
  (let ((x "foo") (y "bar"))
    (osta-html `(:p:id.class (@ :another-attr x) y)))
that evaluate to:
  "<p id=\"id\" class=\"class\" another-attr=\"foo\">bar</p>"