fix wrong parsing when using `one-build-only-html` in async call
To parse correctly a buffer with `org-element-parse-buffer`, we must
have call `org-mode` in that buffer before doing the parsing.  This
wasn't the case with the previous implementation in async call.

Here is an example of that problem with a bold element starting a line
that is parsed as a headline if we don't call `org-mode` first:

(with-temp-buffer
  (require 'org-element)
  (insert "* foo\n*Algebra* and *cryptography*\n* bar")
  (org-element-parse-buffer))
;; (org-data
;;  nil
;;  (headline
;;   (:raw-value "foo" :begin 1 :end 7 ...))
;;  (headline
;;   (:raw-value "Algebra* and *cryptography*" :begin 7 :end 36 ...))
;;  (headline
;;   (:raw-value "bar" :begin 36 :end 41 ...)))